nlb-setup/requirements.txt
aws-cdk-lib==2.47.0
constructs>=10.0.0,<11.0.0
dnspython3>=1.15.0
